Fix lint and remove TypeScript support
