+++ .eslintrc.json
@@ -6,12 +6,9 @@
     "extends": [
         "airbnb",
         "eslint:recommended",
-        "plugin:@typescript-eslint/eslint-recommended",
-        "plugin:@typescript-eslint/recommended",
         "plugin:prettier/recommended",
         "plugin:react/recommended"
     ],
-    "parser": "@typescript-eslint/parser",
     "parserOptions": {
         "ecmaFeatures": {
             "jsx": true
@@ -20,7 +17,6 @@
         "sourceType": "module"
     },
     "plugins": [
-        "@typescript-eslint",
         "prettier",
         "react"
     ],
