--- src/App.js
@@ -1,8 +1,0 @@
-function App() {
-  return (
-    <div className="App">
-    </div>
-  );
-}
-
-export default App;
