+++ src/App.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function App() {
+  return <div className="App" />;
+}
+
+export default App;
